requirements.txt
aiofiles==22.1.0
aiohttp==3.8.1
aiosignal==1.2.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.0
clean-text==0.6.0
cleantext==1.1.4
click==8.1.3
colorama==0.4.5
decorator==5.1.1
Deprecated==1.2.13
discord-py-interactions @ git+https://github.com/interactions-py/library.git@fa03bc2b74a2502ddaf4502ab055241861f81ed1
emoji==1.7.0
enhanced==4.0.0
et-xmlfile==1.1.0
ff3==1.0.1
Flask==2.2.1
frozenlist==1.3.1
ftfy==6.1.1
h11==0.13.0
humanfriendly==10.0
idna==3.3
importlib-metadata==4.12.0
install==1.3.5
interactions-autosharder==2.1.0
interactions-database @ git+https://github.com/axiinyaa/interactions-database@8bfd062194d77a6655b20f03d925dd94dbc2bb90
interactions-files==1.1.5
interactions-lavalink @ git+https://github.com/interactions-py/lavalink.git@22451cd33460733cba5aa7b61e82f778bf877b3f
interactions-persistence==2.2.2
interactions-tasks==1.0.0
interactions-wait-for==1.0.6
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
lavalink==4.0.3
lyricsgenius==3.0.1
MarkupSafe==2.1.1
multidict==6.0.2
mysql-connector-python==8.0.31
nltk==3.7
numpy==1.23.1
openai==0.23.0
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.5.0
pandas-stubs==1.4.4.220919
phantomjs==1.4.1
Pillow==9.2.0
protobuf==3.20.1
pycryptodome==3.15.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytube==12.1.0
pytz==2022.2.1
redis==4.3.4
regex==2022.9.13
replit==3.2.4
requests==2.28.1
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spotipy==2.20.0
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
types-pytz==2022.2.1.0
typing-extensions==3.10.0.2
urllib3==1.26.12
validators==0.20.0
wcwidth==0.2.5
Werkzeug==2.2.1
wrapt==1.14.1
wsproto==1.2.0
yarl==1.8.1
youtube.py==2.1.0
zipp==3.8.1
